+++ Code/jbigkit-2.1/libjbig/jbig85.c
@@ -902,7 +902,7 @@ int jbg85_dec_in(struct jbg85_dec_state *s, unsigned char *data, size_t len,
 	*cnt += s->comment_skip;
 	s->comment_skip = 0;
       } else {
-	s->comment_skip -= len - *cnt;
+	s->comment_skip -= (unsigned long) (len - *cnt);
 	*cnt = len;
       }
       continue;
